LayoutTests/platform/chromium/accessibility/webkit-scrollarea-position-expected.txt

text

text
This tests that the WebKit generated scroll areas are correct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scroll.role is 'AXRole: AXScrollArea'
PASS web.role is 'AXRole: AXWebArea'
PASS scroll.y < web.y is true
PASS web.y > initialWebY is true
PASS successfullyParsed is true

TEST COMPLETE
